+++ changelog.md
@@ -7,6 +7,7 @@ All Notable changes to Packager will be documented in this file.
 ### Added
 - The `new` command now also accepts an option `--i` To interactively make a package and change all Skeleton placholders.
 - Composer autoloads are dumped after installing or creating a package.
+- The `packager:check` function to check the composer lockfile for security vulnerabilities.
 
 ### Fixed
 - Replacing of the Skeleton placeholders.